+++ src/erc20-minter/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("{}", erc20_minter::idl());
+}
