+++ src/erc20-minter/src/state/config.rs
@@ -12,7 +12,7 @@ use super::Settings;
 use crate::client::EvmLink;
 use crate::memory::{CONFIG_MEMORY_ID, MEMORY_MANAGER};
 
-/// Configuration storage for the evm-minter canister.
+/// Configuration storage for the erc20-minter canister.
 pub struct Config {
     data: StableCell<ConfigData, VirtualMemory<DefaultMemoryImpl>>,
 }
