--- src/evm-minter/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("{}", evm_minter::idl());
-}
