+++ src/icrc2-minter/src/build_data.rs
@@ -46,7 +46,7 @@ mod tests {
     fn should_create_build_data() {
         let build_data = canister_build_data();
 
-        assert_eq!(build_data.pkg_name, "minter_canister");
+        assert_eq!(build_data.pkg_name, "icrc2-minter");
 
         assert_eq!(build_data.cargo_target_triple, CARGO_TARGET_TRIPLE);
         assert_eq!(build_data.cargo_features, CARGO_FEATURES);
