+++ src/icrc2-minter/src/canister/inspect.rs
