+++ src/icrc2-minter/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("{}", icrc2_minter::idl());
+}
