+++ src/integration-tests/Cargo.toml
@@ -21,7 +21,7 @@ ic-exports = { workspace = true, features = ["icrc", "ledger"] }
 ic-helpers = { workspace = true }
 ic-log = { workspace = true }
 ic-stable-structures = { workspace = true }
-minter_canister = { path = "../minter-canister" }
+icrc2-minter = { path = "../icrc2-minter" }
 minter-client = { workspace = true }
 minter-contract-utils = { path = "../minter-contract-utils", features = ["test-contracts"] }
 minter-did = { workspace = true }
@@ -39,7 +39,7 @@ ic-test-utils = { workspace = true }
 rlp = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["fs", "process"] }
-evm-minter = { path = "../evm-minter" }
+erc20-minter = { path = "../erc20-minter" }
 signature-verification-canister-client = { workspace = true }
 
 [features]
