+++ src/integration-tests/tests/utils/wasm.rs
@@ -34,9 +34,9 @@ pub async fn get_spender_canister_bytecode() -> Vec<u8> {
 }
 
 /// Returns the bytecode of the minter evm
-pub async fn get_evm_minter_canister_bytecode() -> Vec<u8> {
+pub async fn get_erc20_minter_canister_bytecode() -> Vec<u8> {
     static CANISTER_BYTECODE: OnceCell<Vec<u8>> = OnceCell::new();
-    get_or_load_wasm(&CANISTER_BYTECODE, "evm-minter.wasm.gz").await
+    get_or_load_wasm(&CANISTER_BYTECODE, "erc20-minter.wasm.gz").await
 }
 
 /// Returns the bytecode of the evmc canister - Testnet
@@ -48,7 +48,7 @@ pub async fn get_evm_testnet_canister_bytecode() -> Vec<u8> {
 /// Returns the bytecode of the minter canister
 pub async fn get_minter_canister_bytecode() -> Vec<u8> {
     static CANISTER_BYTECODE: OnceCell<Vec<u8>> = OnceCell::new();
-    get_or_load_wasm(&CANISTER_BYTECODE, "minter.wasm.gz").await
+    get_or_load_wasm(&CANISTER_BYTECODE, "icrc2-minter.wasm.gz").await
 }
 
 async fn load_wasm_bytecode_or_panic(wasm_name: &str) -> Vec<u8> {
