--- src/minter-canister/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("{}", minter_canister::idl());
-}
